Finalyearproject-main/med-detector/backend/requirements.txt
Flask
Flask-Cors
PyMongo
python-dotenv
PyJWT
bcrypt